pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "grounding"
version = "0.1.0"
description = "RAG pipeline with Voyage AI embeddings and Qdrant vector storage"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "qdrant-client>=1.12.0",
    "voyageai>=0.3.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "fastembed>=0.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/grounding"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

